+++ package.json
@@ -1,12 +1,16 @@
 {
     "name": "matrix-cypher",
-    "version": "0.1.3",
+    "version": "0.1.4",
     "description": "Cypher is a small stateless client for the matrix ecosystem",
     "author": "Jorik Schellekens <joriksch@gmail.com>",
     "license": "Apache-2.0",
-    "main": "lib/index.js",
-    "typings": "lib/index.d.ts",
+    "main": "lib/matrix-cypher.umd.js",
+    "module": "lib/matrix-cypher.es5.js",
+    "typings": "lib/types/matrix-cypher.d.ts",
     "devDependencies": {
+        "@rollup/plugin-commonjs": "^14.0.0",
+        "@rollup/plugin-json": "^4.1.0",
+        "@rollup/plugin-node-resolve": "^8.4.0",
         "@types/jest": "^26.0.9",
         "eslint": "^7.6.0",
         "eslint-config-matrix-org": "^0.1.1",
@@ -14,12 +18,15 @@
         "jest": "^26.2.2",
         "lint-staged": "^10.2.11",
         "prettier": "^2.0.5",
+        "rollup": "^2.23.1",
+        "rollup-plugin-sourcemaps": "^0.6.2",
+        "rollup-plugin-typescript2": "^0.27.2",
         "ts-jest": "^26.1.4",
         "typescript": "^3.9.5"
     },
     "scripts": {
-        "build": "tsc",
-        "watch": "tsc --watch",
+        "build": "tsc --module commonjs && rollup -c rollup.config.ts",
+        "start": "rollup -c rollup.config.ts -w",
         "lint": "eslint src"
     },
     "husky": {
@@ -42,6 +49,6 @@
     "jest": {
         "preset": "ts-jest",
         "testEnvironment": "node",
-        "rootDir": "src"
+        "rootDir": "test"
     }
 }
